requirements-offline.txt
requests
pybase64
firebase-admin
requests-toolbelt
Flask
jsonpickle
pillow
numpy
opencv-python==4.4.0.46
boto3
simplejson